fix: revert fp16 to bf16 — T4 emulates bf16 in software, fp16 causes NaN in gradients

+++ train.py
@@ -178,10 +178,10 @@ class GPT(nn.Module):
         head_dim = self.config.n_embd // self.config.n_head
         cos, sin = self._precompute_rotary_embeddings(self.rotary_seq_len, head_dim)
         self.cos, self.sin = cos, sin
-        # Cast embeddings to fp16 (T4 Turing has no native bf16)
-        self.transformer.wte.to(dtype=torch.float16)
+        # Cast embeddings to bf16 (T4 emulates bf16 — slower but stable, fp16 causes NaN)
+        self.transformer.wte.to(dtype=torch.bfloat16)
         for ve in self.value_embeds.values():
-            ve.to(dtype=torch.float16)
+            ve.to(dtype=torch.bfloat16)
 
     def _precompute_rotary_embeddings(self, seq_len, head_dim, base=10000, device=None):
         if device is None:
@@ -191,7 +191,7 @@ class GPT(nn.Module):
         t = torch.arange(seq_len, dtype=torch.float32, device=device)
         freqs = torch.outer(t, inv_freq)
         cos, sin = freqs.cos(), freqs.sin()
-        cos, sin = cos.half(), sin.half()
+        cos, sin = cos.bfloat16(), sin.bfloat16()
         cos, sin = cos[None, :, None, :], sin[None, :, None, :]
         return cos, sin
 
@@ -445,7 +445,7 @@ torch.manual_seed(42)
 torch.cuda.manual_seed(42)
 torch.set_float32_matmul_precision("high")
 device = torch.device("cuda")
-autocast_ctx = torch.amp.autocast(device_type="cuda", dtype=torch.float16)
+autocast_ctx = torch.amp.autocast(device_type="cuda", dtype=torch.bfloat16)  # T4 emulates bf16 in software — slower but numerically stable (fp16 causes NaN)
 T4_FP16_PEAK_FLOPS = 65e12
 
 tokenizer = Tokenizer.from_directory()
